Added SNAPSTACK_TEST_INSTALL env variable.
Replaced @skip with @skipUnless for the full snapstack run test.

Allows us to trigger a full snapstack test run with an environment
variable, rather than via editing the test file, which is much nicer,
and less vulnerable to accidental checkins of tools that make changes
to a machine outside of the venv.